Store dac and fi in asm

+++ src/test/java/dk/tbsalling/aismessages/ais/messages/BinaryBroadcastMessageTest.java
@@ -66,8 +66,8 @@ public class BinaryBroadcastMessageTest {
 
         InlandShipStaticAndVoyageRelatedData asm = (InlandShipStaticAndVoyageRelatedData) binaryBroadcastMessage.getApplicationSpecificMessage();
 
-        assertEquals(200, asm.getDesignatedAreaCode().intValue());
-        assertEquals(10, asm.getFunctionalId().intValue());
+        assertEquals(200, asm.getDesignatedAreaCode());
+        assertEquals(10, asm.getFunctionalId());
         assertEquals("02325170", asm.getUniqueEuropeanVesselIdentificationNumber());
         assertEquals(Float.valueOf(80.0f), asm.getLengthOfShip());
         assertEquals(Float.valueOf(8.2f), asm.getBeamOfShip());
